remove objc ExceptionWrapper
