+++ Mixpanel/Mixpanel.h
@@ -7,7 +7,6 @@
 //
 
 @import Foundation;
-#import "ExceptionWrapper.h"
 
 //! Project version number for Mixpanel.
 FOUNDATION_EXPORT double MixpanelVersionNumber;
